Use separate files for intents and disambiguation
---
